FunctionApps/FunctionApps/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.27.1
azure-servicebus
twilio==7.9.2
matplotlib==3.5.1